requirements.txt
jupyter
matplotlib
numpy
pandas
scikit-learn
scikit-image
Pillow
scipy
torch
torchvision
libopencv
pyvips
tqdm
img2vec_pytorch
